+++ schedule/core/src/main/kotlin/vp/togedo/service/KafkaService.kt
@@ -2,8 +2,8 @@ package vp.togedo.service
 
 import org.bson.types.ObjectId
 import reactor.core.publisher.Mono
-import vp.togedo.data.dao.GroupSchedule.GroupDao
-import vp.togedo.data.dao.Group.GroupScheduleDao
+import vp.togedo.data.dao.group.GroupDao
+import vp.togedo.data.dao.groupSchedule.GroupScheduleDao
 
 interface KafkaService {
